+++ src/main/java/com/example/system_demo/filter/FilterEncoding.java
@@ -1,0 +1,28 @@
+package com.example.system_demo.filter;
+
+import javax.servlet.*;
+import javax.servlet.annotation.*;
+import java.io.IOException;
+
+@WebFilter(filterName = "FilterEncoding", urlPatterns = "/*")
+public class FilterEncoding implements Filter {
+
+    private String encoding;
+
+    public void init(FilterConfig config) throws ServletException {
+        encoding = config.getInitParameter("encoding");
+        if (encoding == null){
+            encoding = "UTF-8";
+        }
+    }
+
+    public void destroy() {
+        encoding = null;
+    }
+
+    @Override
+    public void doFilter(ServletRequest request, ServletResponse response, FilterChain chain) throws ServletException, IOException {
+        request.setCharacterEncoding(encoding);
+        chain.doFilter(request, response);
+    }
+}
