+++ src/main/java/com/example/system_demo/repository/repository_eval.java
@@ -2,12 +2,11 @@ package com.example.system_demo.repository;
 
 import com.example.system_demo.util.util;
 
-import java.sql.Connection;
-import java.sql.PreparedStatement;
-import java.sql.ResultSet;
-import java.sql.SQLException;
+import java.sql.*;
 import java.util.Arrays;
 import java.util.List;
+
+import static com.example.system_demo.util.UTIL_CONSTANTS.*;
 
 public class repository_eval {
 
@@ -46,6 +45,46 @@ public class repository_eval {
                 people/num_rows
         );
 
+    }
 
+    public static Double Calculator_service_evaluate(int serviceID) throws SQLException {
+        String sql = "select * from service_evaluate where serviceID = ?";
+        Connection connection = util.initConnection();
+        PreparedStatement preparedStatement = connection.prepareStatement(sql);
+        preparedStatement.setInt(1, serviceID);
+        ResultSet resultSet = preparedStatement.executeQuery();
+
+        int num_rows = 0;
+        int sum = 0;
+
+        while (resultSet.next()){
+            num_rows++;
+
+            ResultSetMetaData resultSetMetaData = resultSet.getMetaData();
+            int columnCount = resultSetMetaData.getColumnCount();
+            for (int i = 1; i<= columnCount; i++){
+
+                String columnName = resultSetMetaData.getColumnName(i);
+
+                if (!scale_0.contains(columnName)){
+
+                    int score = (int)resultSet.getObject(i);
+
+                    if (scale_1.contains(columnName)) sum += score;
+                    else if (scale_4.contains(columnName)) sum += 4 * score;
+                    else sum += 2 * score;
+
+                }
+            }
+        }
+
+        double score_evaluate = 0;
+        if (num_rows != 0){
+            score_evaluate = (double) sum/num_rows;
+        }
+
+        util.close(connection, preparedStatement, resultSet);
+
+        return score_evaluate;
     }
 }
