+++ src/main/java/com/example/system_demo/repository/repository_score.java
@@ -2,9 +2,12 @@ package com.example.system_demo.repository;
 
 import com.example.system_demo.model.model_score;
 import com.example.system_demo.util.util;
-import com.example.system_demo.util.util_score_calculator;
 
 import java.sql.*;
+
+import static com.example.system_demo.repository.repository_eval.Calculator_service_evaluate;
+import static com.example.system_demo.repository.repository_process.Calculator_service_process;
+import static com.example.system_demo.repository.repository_service.Calculator_service_info;
 
 public class repository_score {
 
@@ -13,7 +16,7 @@ public class repository_score {
         model_score score = null;
 
         try {
-            util_score_calculator.updateServiceScore(serviceID);
+            updateServiceScore(serviceID);
         } catch (SQLException e) {
             e.printStackTrace();
         }
@@ -37,5 +40,52 @@ public class repository_score {
             e.printStackTrace();
         }
         return null;
+    }
+
+    public static void updateServiceScore(int serviceID) throws SQLException {
+
+        double score_info = Calculator_service_info(serviceID);
+        double score_proc = Calculator_service_process(serviceID);
+        double score_eval = Calculator_service_evaluate(serviceID);
+
+        double avg = (score_info + score_eval + score_proc)/3;
+
+        String sql = "select * from service_score where serviceID = ?";
+        Connection connection = util.initConnection();
+        PreparedStatement preparedStatement = connection.prepareStatement(sql);
+        preparedStatement.setInt(1, serviceID);
+        ResultSet resultSet = preparedStatement.executeQuery();
+
+        Connection connection_update = util.initConnection();
+        PreparedStatement preparedStatement_update;
+
+        if (resultSet.next()){
+
+            String sql_update =
+                    "update service_score set sumAvg = ?, sumAvg1 = ?, sumAvg2 = ?, sumAvg3 = ? where serviceID = ?";
+            preparedStatement_update = connection_update.prepareStatement(sql_update);
+            preparedStatement_update.setDouble(1, avg);
+            preparedStatement_update.setDouble(2, score_info);
+            preparedStatement_update.setDouble(3, score_proc);
+            preparedStatement_update.setDouble(4, score_eval);
+            preparedStatement_update.setInt(5, serviceID);
+
+        }else {
+
+            String sql_insert =
+                    "insert into service_score (serviceID, sumAvg, sumAvg1, sumAvg2, sumAvg3) values (?, ?, ?, ?, ?)";
+            preparedStatement_update = connection_update.prepareStatement(sql_insert);
+            preparedStatement_update.setInt(1, serviceID);
+            preparedStatement_update.setDouble(2, avg);
+            preparedStatement_update.setDouble(3, score_info);
+            preparedStatement_update.setDouble(4, score_proc);
+            preparedStatement_update.setDouble(5, score_eval);
+
+        }
+        preparedStatement_update.executeUpdate();
+
+        util.close(connection, preparedStatement, resultSet);
+        util.close(connection_update, preparedStatement_update, null);
+
     }
 }
